Port to Qt 5.15 and Apple Silicon (arm64)
Qt 4.8 cannot be built on arm64 Macs, so the client now builds against
Qt 5.15 (Homebrew qt@5) and liblastfm's Qt5 build (liblastfm5), producing
a native Apple Silicon app.

Build system:
- link liblastfm5 from a parallel liblastfm checkout (LIBLASTFM_ROOT,
  default ../liblastfm/_install)
- @rpath install names for our dylibs; apps get rpaths for dev runs and
  macdeployqt packaging. The legacy bundleFrameworks.sh post-link step is
  gone (it predates macdeployqt and corrupted install names)
- deployment target 10.6 -> 11.0
- twiddly (iPodScrobbler helper) builds again on mac; fingerprinter
  remains disabled

Code, mechanical Qt5 fixes:
- Q_WS_MAC -> Q_OS_MAC, toAscii -> toLatin1, QUrl query API -> QUrlQuery,
  qInstallMsgHandler -> qInstallMessageHandler, setResizeMode ->
  setSectionResizeMode, missing includes, QVariant/QUuid strictness

macOS integration:
- analytics service disabled: it needed QtWebKit (dead in Qt5) and only
  fed the long-gone Universal Analytics endpoint
- Sparkle 2.x supported when Sparkle.framework is installed (HAVE_SPARKLE),
  no-op otherwise; Growl likewise optional (HAVE_GROWL) and effectively
  retired in favour of NSUserNotification
- menu bar icon is now a Retina-ready template image (fixes half-size
  icon; adapts to dark menu bars; dimmed glyph = scrobbling off)
- force Aqua appearance: the stylesheet is a light-only design and Qt5
  otherwise picks up the dark system palette, making panes unreadable
- fix 100% CPU spin: prettyTime scheduled a 0ms repeating refresh for
  tracks with invalid timestamps (e.g. friends' now-playing rows)
- Enable SSL now defaults to on

+++ lib/unicorn/widgets/LfmListViewWidget.cpp
@@ -176,7 +176,7 @@ LfmListModel::data( const QModelIndex & index, int role ) const
             return item.m_type->toString();
 
         case CursorRole:
-            return Qt::PointingHandCursor;
+            return static_cast<int>( Qt::PointingHandCursor );
 
         case WwwRole:
             return item.m_type->www();
